@@ -27,11 +27,11 @@
     "url": "https://github.com/Ilhasoft/unnnic.git"
   },
   "dependencies": {
-    "@vue/compat": "^3.3.9",
-    "@vue/compiler-sfc": "^3.3.9",
+    "@vue/compat": "^3.4.8",
+    "@vue/compiler-sfc": "^3.4.8",
     "@vueuse/components": "^10.4.1",
     "moment": "^2.30.1",
-    "vue": "^3.3.9",
+    "vue": "^3.4.8",
     "vue-i18n": "9",
     "vue-the-mask": "^0.11.1"
   },
